pyproject.toml
[project]
name = "group-av"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=23.0.0",
    "folium>=0.15.0",
    "fpdf2>=2.7.4",
    "geopy>=2.4.0",
    "googlemaps>=4.10.0",
    "groq>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-groq>=0.1.0",
    "numpy>=1.24.0",
    "openai-whisper>=20230918",
    "pyaudio>=0.2.11",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.31.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "streamlit>=1.31.0",
    "torch>=2.0.0",
    "uvicorn>=0.40.0",
]
